// src/types/graphql.tsx
export interface Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
}
// ...
export interface File extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  childAuthorJson: AuthorJson | null;
  childImageSharp: ImageSharp | null;
  childMarkdownRemark: MarkdownRemark | null;
  internal: internal_12 | null;
  sourceInstanceName: string | null;
  absolutePath: string | null;
  relativePath: string | null;
  extension: string | null;
  size: string | null;
  prettySize: string | null;
  modifiedTime: string | null;
  accessTime: string | null;
  changeTime: string | null;
  birthTime: string | null;
  root: string | null;
  dir: string | null;
  base: string | null;
  ext: string | null;
  name: string | null;
  relativeDirectory: string | null;
  dev: number | null;
  mode: number | null;
  nlink: number | null;
  uid: number | null;
  gid: number | null;
  rdev: number | null;
  blksize: string | null;
  ino: number | null;
  blocks: number | null;
  atimeMs: number | null;
  mtimeMs: number | null;
  ctimeMs: number | null;
  birthtimeMs: number | null;
  atime: string | null;
  mtime: string | null;
  ctime: string | null;
  birthtime: string | null;
}
// ...
export interface AuthorJson extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  bio: string | null;
  avatar: File | null;
  twitter: string | null;
  github: string | null;
  internal: internal_13 | null;
}

export interface internal_13 {
  contentDigest: string | null;
  type: string | null;
  owner: string | null;
}

export interface ImageSharp extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  internal: internal_14 | null;
  original: ImageSharpOriginal | null;
  resolutions: ImageSharpResolutions | null;
  sizes: ImageSharpSizes | null;
  fluid: ImageSharpfluid | null;
  fixed: ImageSharpfixed | null;
  responsiveSizes: ImageSharpResponsiveSizes | null;
  resize: ImageSharpResize | null;
}
// ...
export interface internal_14 {
  contentDigest: string | null;
  type: string | null;
  owner: string | null;
}

export interface ImageSharpOriginal {
  width: number | null;
  height: number | null;
  src: string | null;
}
// ...
export interface ImageSharpResolutions {
  base64: string | null;
  tracedSVG: string | null;
  aspectRatio: number | null;
  width: number | null;
  height: number | null;
  src: string | null;
  srcSet: string | null;
  srcWebp: string | null;
  srcSetWebp: string | null;
  originalName: string | null;
}

export interface ImageSharpSizes {
  base64: string | null;
  tracedSVG: string | null;
  aspectRatio: number | null;
  src: string | null;
  srcSet: string | null;
  srcWebp: string | null;
  srcSetWebp: string | null;
  sizes: string | null;
  originalImg: string | null;
  originalName: string | null;
}

export interface ImageSharpfluid {
  base64: string | null;
  aspectRatio: number | null;
  width: number | null;
  height: number | null;
  src: string | null;
  srcSet: string | null;
  originalName: string | null;
}

export interface ImageSharpfixed {
  base64: string | null;
  aspectRatio: number | null;
  width: number | null;
  height: number | null;
  src: string | null;
  srcSet: string | null;
  originalName: string | null;
}

export interface ImageSharpResponsiveSizes {
  base64: string | null;
  aspectRatio: number | null;
  src: string | null;
  srcSet: string | null;
  sizes: string | null;
  originalImg: string | null;
  originalName: string | null;
}

export interface ImageSharpResize {
  src: string | null;
  tracedSVG: string | null;
  width: number | null;
  height: number | null;
  aspectRatio: number | null;
  originalName: string | null;
}

export interface MarkdownRemark extends Node {
  id: string;
  parent: Node | null;
  children: Array<Node> | null;
  internal: internal_15 | null;
  frontmatter: frontmatter_2 | null;
  fileAbsolutePath: string | null;
  fields: fields_2 | null;
  html: string | null;
  excerpt: string | null;
  headings: Array<MarkdownHeading> | null;
  timeToRead: number | null;
  tableOfContents: string | null;
  wordCount: wordCount | null;
}

export interface ExcerptMarkdownRemarkArgs {
  pruneLength: number | null;
}

export interface HeadingsMarkdownRemarkArgs {
  depth: HeadingLevels | null;
}

export interface internal_15 {
  content: string | null;
  contentDigest: string | null;
  type: string | null;
  owner: string | null;
  fieldOwners: fieldOwners_2 | null;
}

export interface fieldOwners_2 {
  slug: string | null;
}

export interface frontmatter_2 {
  title: string | null;
  createdDate: string | null;
  updatedDate: string | null;
  author: AuthorJson | null;
  tags: Array<string> | null;
  image: File | null;
  draft: boolean | null;
  _PARENT: string | null;
  parent: string | null;
}

export interface CreatedDatefrontmatter_2Args {
  formatString: string | null;
  fromNow: boolean | null;
  difference: string | null;
  locale: string | null;
}

export interface UpdatedDatefrontmatter_2Args {
  formatString: string | null;
  fromNow: boolean | null;
  difference: string | null;
  locale: string | null;
}

export interface fields_2 {
  slug: string | null;
}

export type HeadingLevels = "h1" | "h2" | "h3" | "h4" | "h5" | "h6";

export interface MarkdownHeading {
  value: string | null;
  depth: number | null;
}

export interface wordCount {
  paragraphs: number | null;
  sentences: number | null;
  words: number | null;
}

export interface internal_12 {
  contentDigest: string | null;
  mediaType: string | null;
  type: string | null;
  owner: string | null;
}